removed unused files. consolidated tests.
