+++ example.py
@@ -3,16 +3,16 @@
 import rvrbase as rvrbase
 
 # Log something.
-thelogger = rvrbase.Rvrlogger()
-thelogger.log_application_event(
-    type='warning', message='No worries, just testing here.')
+newdelegate = rvrbase.Mydelegate(rvrbase.CONFIG_FILE)
+# logs to both the system (using logger) and Azure.
+newdelegate.log_app_event(type='warning', message='Dit is een test met nieuw design pattern.')
+# Sends the message to Azure only.
+newdelegate.send_az_app_event(type='info', message='Just relax.')
+# sends a metric to Azure only.
+newdelegate.send_az_metric(log_type='info', metric_name='somemetric', value=5)
+
+# Send a notification to the browser(s)
+newdelegate.send_browser_notification(type='info', message='No worries, just testing here.')
 
 # print something from the configuration file.
-rvrconfig = rvrbase.Rvrconfig(rvrbase.CONFIG_FILE)
-print(rvrconfig.q1('hwip'))
-
-newdelegate = rvrbase.Mydelegate(rvrbase.CONFIG_FILE)
 print(newdelegate.q1('hwip'))
-newdelegate.send_browser_notification(type='info', message='Dit is een test met nieuw design pattern.')
-newdelegate.send_az_app_event(type='info', message='Dit is een test met nieuw design pattern.')
-newdelegate.send_az_metric(log_type='info', metric_name='somemetric', value=5)
