+++ src/rvrbase/__init__.py
@@ -1,6 +1,4 @@
-from rvrbase.rvrlogger_api import * # noqa
 from rvrbase.rvrconfig_delegate import * # noqa
-from rvrbase.rvrlogger_delegate import * # noqa
 from rvrbase.constants import * # noqa
 from rvrbase.__about__ import (__version__,)
 from rvrbase.delegate import * # noqa
