+++ src/rvrbase/delegate.py
@@ -1,4 +1,5 @@
-﻿from platform import node
+import logging
+from platform import node
 import json
 import sys
 
@@ -22,6 +23,9 @@ class Mydelegate(rvrconfig_delegate.Rvrconfig):
         self.workspace_id = self.q1('workspace_id')
         self.workspace_prim_key = self.q1('primary_key')
 
+        message = f'Starting, notify URL: {self.notifcation_api.notify.endpoint}' # todo to constant.
+        self.send_az_app_event(type='info', message=message)
+
     def send_browser_notification(self, message, type):
         if type not in VALID_MESSAGE_TYPES:
             raise ValidationException(
@@ -29,6 +33,25 @@ class Mydelegate(rvrconfig_delegate.Rvrconfig):
 
         self.notifcation_api.send_browser_notification(
             NOTIFY_APPLICATION_EVENT.format(type=type, node=node(), message=message))
+
+    def log_app_event(self, type, message, notify_message=False):
+        if type not in VALID_MESSAGE_TYPES:
+            raise ValidationException(
+                f'{type} is not in {VALID_MESSAGE_TYPES}.')  # todo use constants.
+                
+        if type == 'debug':
+            logging.debug(message)
+        elif type == 'info':
+            logging.info(message)
+        elif type == 'warning':
+            logging.warning(message)
+        elif type == 'error':
+            logging.error(message)
+
+        self.send_az_app_event(type, message)
+
+        if notify_message:
+            self.notifcation_api.send_browser_notification(message=message)
 
     def send_az_app_event(self, type, message):
         if type not in VALID_MESSAGE_TYPES:
