+++ src/test/testdata/config.ini
@@ -1,3 +1,3 @@
 [AZ_LOG_ANALYTICS_WORKSPACE]
-workspace_id = some_id
-primary_key = somekey
+workspace_id = 7756814b-7720-4b6d-9fb6-0aa03fe97658
+primary_key = OmCS6OxSmvid7GL1l48w4onIkNclgI37Z2QMuA12hXPXN7YsLLxcug866K5ZW5zLFtbGzEcOPhPnhBqSbMEe/7==
